requirements.txt

# Google Sheets API v4 client (read/write HITL dashboard rows)
google-api-python-client>=2.100.0

# Service-account authentication for Sheets API
google-auth>=2.20.0

# Async web framework (webhook receiver, ADA chatbot API)
fastapi>=0.115.0

# ASGI server for FastAPI
uvicorn[standard]>=0.30.0

# Load .env configuration at startup
python-dotenv>=1.0.0

# Request/response validation and serialization
pydantic>=2.0.0
